Creating useOfflineData and deleting OfflineUtil
